Tests/LibWeb/Text/expected/WebAudio/BaseAudioContext-decodeAudioData-mp3.txt

channels: 2
sampleRate: 44100
length: 655488
